+++ native/comrak_nif/src/lib.rs
@@ -325,6 +325,5 @@ fn do_safe_html(
 
 #[rustler::nif(schedule = "DirtyCpu")]
 pub fn text_to_anchor(env: Env<'_>, text: String) -> String {
-    let mut anchorizer = Anchorizer::new();
-    anchorizer.anchorize(text)
+    Anchorizer::new().anchorize(text)
 }